backend/requirements.txt
asgirefnpx 
Django
django-cors-headers
django-dotenv
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
tzdata